cutmylink/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==4.0.0
django-rename-app==0.1.6
django-user-agents==0.4.0
djangorestframework==3.14.0
dnspython==2.0.0
idna==3.4
ipwhois==1.2.0
psycopg2-binary==2.9.6
pypng==0.20220715.0
pyTelegramBotAPI==4.10.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.6.1
ua-parser==0.16.1
urllib3==1.26.15
user-agents==2.2.0
